[#205, #225]refactor: UserInfoResponse 위치 변경
