+++ src/main/java/com/runninghi/feedback/command/infrastructure/service/ApiFeedbackCommandInfraService.java
@@ -3,7 +3,7 @@ package com.runninghi.feedback.command.infrastructure.service;
 import com.runninghi.common.annotation.InfraService;
 import com.runninghi.feedback.command.application.dto.response.FeedbackUserResponse;
 import com.runninghi.feedback.command.domain.service.ApiFeedbackCommandDomainService;
-import com.runninghi.user.command.application.dto.user.response.UserInfoResponse;
+import com.runninghi.user.query.application.dto.user.response.UserInfoResponse;
 import com.runninghi.user.query.application.service.UserQueryService;
 import lombok.RequiredArgsConstructor;
