+++ src/main/java/com/runninghi/feedback/query/infrastructure/service/ApiFeedbackInfraQueryDomainService.java
@@ -3,7 +3,7 @@ package com.runninghi.feedback.query.infrastructure.service;
 import com.runninghi.common.annotation.InfraService;
 import com.runninghi.feedback.query.application.dto.response.FeedbackUserResponse;
 import com.runninghi.feedback.query.domain.service.ApiFeedbackQueryDomainService;
-import com.runninghi.user.command.application.dto.user.response.UserInfoResponse;
+import com.runninghi.user.query.application.dto.user.response.UserInfoResponse;
 import com.runninghi.user.query.application.service.UserQueryService;
 import lombok.RequiredArgsConstructor;
