+++ src/main/java/com/runninghi/user/command/application/controller/SignCommandController.java
@@ -23,7 +23,7 @@ public class SignCommandController {
 
     /* 아이디 중복 확인 API */
     @Operation(summary = "아이디 중복 확인")
-    @RequestMapping("/api/v1/verifyDuplicationId")
+    @PostMapping("/api/v1/verifyDuplicationId")
     public ApiResponse verifyDuplicationId(@RequestBody VerifyDuplicationIdRequest request) {
         return ApiResponse.success("중복 확인 결과가 나왔습니다.", signCommandService.verifyDuplicationId(request));
     }
