--- src/main/java/com/runninghi/user/query/application/dto/test.java
@@ -1,4 +1,0 @@
-package com.runninghi.user.query.application.dto;
-
-public class test {
-}
