+++ src/main/java/com/runninghi/user/query/application/service/UserQueryService.java
@@ -2,8 +2,8 @@ package com.runninghi.user.query.application.service;
 
 import com.runninghi.user.command.application.dto.sign_in.request.SignInRequest;
 import com.runninghi.user.command.application.dto.user.RefreshTokenDTO;
-import com.runninghi.user.command.application.dto.user.response.UserInfoResponse;
 import com.runninghi.user.command.domain.aggregate.entity.User;
+import com.runninghi.user.query.application.dto.user.response.UserInfoResponse;
 import com.runninghi.user.query.infrastructure.repository.UserQueryRefreshTokenRepository;
 import com.runninghi.user.query.infrastructure.repository.UserQueryRepository;
 import lombok.RequiredArgsConstructor;
@@ -19,7 +19,7 @@ import java.util.UUID;
 public class UserQueryService {
     private final UserQueryRepository userQueryRepository;
     private final UserQueryRefreshTokenRepository userQueryRefreshTokenRepository;
-    
+
     // 회원 정보 조회
     public UserInfoResponse findUserInfo(UUID id) {
         return userQueryRepository.findById(id)
