+++ src/test/java/com/runninghi/user/query/application/service/AdminQueryServiceTest.java
@@ -1,10 +1,10 @@
 package com.runninghi.user.query.application.service;
 
 
-import com.runninghi.user.command.application.dto.user.response.UserInfoResponse;
 import com.runninghi.user.command.domain.aggregate.entity.User;
 import com.runninghi.user.command.domain.aggregate.entity.enumtype.Role;
 import com.runninghi.user.command.domain.repository.UserCommandRepository;
+import com.runninghi.user.query.application.dto.user.response.UserInfoResponse;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
